requirement.txt
python-dotenv
pytest
pycryptodome
pytest-mock